added health system and apply_damage event

+++ src/components/health.rs
@@ -1,11 +1,11 @@
 pub struct Health {
-    pub amount:i32
+    pub amount:f32
 }
 
 impl Default for Health {
     fn default() -> Self {
         Health { 
-            amount:100
+            amount:100.0
         }
     }
 }
